Rename I19 gda module and fix calls where necessary
