+++ setup.cfg
@@ -51,4 +51,4 @@ console_scripts =
     generate_nexus = nexgen.command_line.nexus_generator:main
     copy_nexus = nexgen.command_line.copy_nexus:main
     nexgen_phil = nexgen.command_line.phil_files_cli:main
-    I19_nexus = nexgen.beamlines.I19_2_nxs:main
+    I19_nexus = nexgen.beamlines.I19_2_cli:gda_writer
